Initial take on socket memory access.
Possibly it should be turned into a parameterised module, but as it's
for internal use only there is probably not much value. Also, it has to
work out the memory areas for all sockets anyway, so the current
approach gives a more accurate representation of what it does I think.

Needs testing; Lots and lots of testing... :)